Numerous changes to better support other OSes
This is primarily FreeBSD focused, but this commit addresses other
things that were hackily/poorly done. For example, some things were not
done simply because OpenBSD was the target OS, not, say, because we
aren't using a dev environment.

This is nearly ready to test on a live FreeBSD box.