Don't run zsh at the end of the install
This didn't allow Chezmoi to finish, and if you forget to exit the new
shell, you're left with a hanging Chezmoi process, and if you kill that,
you have to start the init process all over again.

So, you need to restart your shell, but you miss out on the crummy side
effect. I think it's a net positive change.